// web/components/data/employee-data.d.ts
export declare const genderOptions: {
    value: string;
    label: string;
}[];
export declare const bloodGroupOptions: {
    value: string;
    label: string;
}[];
export declare const maritalStatusOptions: {
    value: string;
    label: string;
}[];
export declare const statusOptionsEmployee: ({
    value: "active";
    label: string;
} | {
    value: "inactive";
    label: string;
})[];